Reorganized unit tests by functional area (#1224)
# Goal
The goal of this PR is to reorganize unit tests by functional area

Closes  #1211
