+++ pallets/messages/src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod mock;
+mod other_tests;
