+++ pallets/msa/src/tests/mod.rs
@@ -1,0 +1,12 @@
+pub mod mock;
+
+mod creation_tests;
+mod delegation_tests;
+mod governance_tests;
+mod other_tests;
+mod public_key_tests;
+mod replay_tests;
+mod retirement_tests;
+mod schema_permission_tests;
+mod signature_registry_tests;
+mod signed_extension_tests;
