+++ pallets/msa/src/tests/other_tests.rs
@@ -1,0 +1,603 @@
+use frame_support::{
+	assert_err, assert_noop, assert_ok,
+	dispatch::{GetDispatchInfo, Pays},
+	BoundedBTreeMap,
+};
+
+use sp_core::{crypto::AccountId32, sr25519, Encode, Pair};
+use sp_runtime::MultiSignature;
+
+use crate::{
+	ensure,
+	tests::mock::*,
+	types::{AddProvider, PermittedDelegationSchemas, EMPTY_FUNCTION},
+	Config, DispatchResult, Error, Event, ProviderToRegistryEntry,
+};
+
+use common_primitives::{
+	msa::{Delegation, DelegatorId, ProviderId, ProviderRegistryEntry, SchemaGrantValidator},
+	node::BlockNumber,
+	schema::{SchemaId, SchemaValidator},
+	utils::wrap_binary_data,
+};
+
+pub fn assert_revoke_delegation_by_delegator_no_effect(
+	test_account: AccountId32,
+	provider_msa_id: u64,
+) {
+	let event_count = System::event_count();
+	assert_ok!(Msa::revoke_delegation_by_delegator(
+		RuntimeOrigin::signed(test_account.clone()),
+		provider_msa_id
+	));
+	assert_eq!(event_count, System::event_count())
+}
+
+#[test]
+pub fn test_get_owner_of() {
+	new_test_ext().execute_with(|| {
+		assert_eq!(Msa::get_owner_of(&test_public(1)), None);
+
+		assert_ok!(Msa::create(test_origin_signed(1)));
+
+		assert_eq!(Msa::get_owner_of(&test_public(1)), Some(1));
+	});
+}
+
+#[test]
+pub fn test_ensure_msa_owner() {
+	new_test_ext().execute_with(|| {
+		assert_noop!(Msa::ensure_msa_owner(&test_public(1), 1), Error::<Test>::NoKeyExists);
+
+		assert_ok!(Msa::add_key(1, &test_public(1), EMPTY_FUNCTION));
+
+		assert_eq!(Msa::ensure_msa_owner(&test_public(1), 1), Ok(()));
+	});
+}
+
+#[test]
+pub fn add_provider_to_msa_is_success() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let provider_account = key_pair.public();
+
+		let (delegator_pair, _) = sr25519::Pair::generate();
+		let delegator_account = delegator_pair.public();
+
+		// Create provider account and get its MSA ID (u64)
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
+		let provider_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
+
+		// Create delegator account and get its MSA ID (u64)
+		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
+		let delegator_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(delegator_account.0)).unwrap();
+
+		// Register provider
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(provider_account.into()),
+			Vec::from("Foo")
+		));
+
+		let (delegator_signature, add_provider_payload) =
+			create_and_sign_add_provider_payload(delegator_pair, provider_msa);
+
+		set_schema_count::<Test>(10);
+
+		assert_ok!(Msa::grant_delegation(
+			RuntimeOrigin::signed(provider_account.into()),
+			delegator_account.into(),
+			delegator_signature,
+			add_provider_payload
+		));
+
+		let provider = ProviderId(provider_msa);
+		let delegator = DelegatorId(delegator_msa);
+
+		assert_eq!(
+			Msa::get_delegation(delegator, provider),
+			Some(Delegation { revoked_at: 0, schema_permissions: Default::default() })
+		);
+
+		System::assert_last_event(
+			Event::DelegationGranted {
+				delegator_id: delegator_msa.into(),
+				provider_id: provider_msa.into(),
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+pub fn ensure_valid_msa_key_is_successfull() {
+	new_test_ext().execute_with(|| {
+		assert_noop!(Msa::ensure_valid_msa_key(&test_public(1)), Error::<Test>::NoKeyExists);
+
+		assert_ok!(Msa::create(test_origin_signed(1)));
+
+		assert_ok!(Msa::ensure_valid_msa_key(&test_public(1)));
+	});
+}
+
+#[test]
+pub fn add_key_with_panic_in_on_success_should_revert_everything() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1u64;
+		let key = test_public(msa_id as u8);
+
+		// act
+		assert_noop!(
+			Msa::add_key(msa_id, &key, |new_msa_id| -> DispatchResult {
+				ensure!(new_msa_id != msa_id, Error::<Test>::InvalidSelfRemoval);
+				Ok(())
+			}),
+			Error::<Test>::InvalidSelfRemoval
+		);
+
+		// assert
+		assert_eq!(Msa::get_msa_by_public_key(&key), None);
+
+		// *Temporarily Removed* until https://github.com/LibertyDSNP/frequency/issues/418 is completed
+		// assert_eq!(Msa::get_msa_keys(msa_id).into_inner(), vec![])
+	});
+}
+
+#[test]
+pub fn revoke_delegation_by_delegator_is_successful() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let provider_account = key_pair.public();
+
+		let (delegator_pair, _) = sr25519::Pair::generate();
+		let delegator_account = delegator_pair.public();
+
+		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
+
+		// Register provider
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(provider_account.into()),
+			Vec::from("Foo")
+		));
+
+		let provider_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
+
+		let (delegator_signature, add_provider_payload) =
+			create_and_sign_add_provider_payload(delegator_pair, provider_msa);
+
+		assert_ok!(Msa::grant_delegation(
+			RuntimeOrigin::signed(provider_account.into()),
+			delegator_account.into(),
+			delegator_signature,
+			add_provider_payload
+		));
+
+		assert_ok!(Msa::revoke_delegation_by_delegator(
+			RuntimeOrigin::signed(delegator_account.into()),
+			2
+		));
+
+		System::assert_last_event(
+			Event::DelegationRevoked { delegator_id: 1.into(), provider_id: 2.into() }.into(),
+		);
+	});
+}
+
+#[test]
+pub fn revoke_provider_is_successful() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let provider_account = key_pair.public();
+
+		let (delegator_pair, _) = sr25519::Pair::generate();
+		let delegator_account = delegator_pair.public();
+
+		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
+
+		let provider_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
+
+		let (delegator_signature, add_provider_payload) =
+			create_and_sign_add_provider_payload(delegator_pair, provider_msa);
+
+		// Register provider
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(provider_account.into()),
+			Vec::from("Foo")
+		));
+
+		assert_ok!(Msa::grant_delegation(
+			RuntimeOrigin::signed(provider_account.into()),
+			delegator_account.into(),
+			delegator_signature,
+			add_provider_payload
+		));
+
+		let delegator_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(delegator_account.0)).unwrap();
+
+		let provider = ProviderId(provider_msa);
+		let delegator = DelegatorId(delegator_msa);
+
+		assert_ok!(Msa::revoke_provider(provider, delegator));
+
+		assert_eq!(
+			Msa::get_delegation(delegator, provider).unwrap(),
+			Delegation { revoked_at: 1, schema_permissions: Default::default() },
+		);
+	});
+}
+
+#[test]
+fn revoke_delegation_by_delegator_does_nothing_when_no_msa() {
+	new_test_ext()
+		.execute_with(|| assert_revoke_delegation_by_delegator_no_effect(test_public(3), 333u64));
+}
+
+#[test]
+pub fn revoke_delegation_by_delegator_does_nothing_if_only_key_is_revoked() {
+	new_test_ext().execute_with(|| {
+		assert_ok!(Msa::create(RuntimeOrigin::signed(test_public(2))));
+		assert_ok!(Msa::delete_key_for_msa(1, &test_public(2)));
+		assert_revoke_delegation_by_delegator_no_effect(test_public(2), 1u64)
+	})
+}
+
+#[test]
+pub fn revoke_delegation_by_delegator_fails_if_has_msa_but_no_delegation() {
+	new_test_ext().execute_with(|| {
+		assert_ok!(Msa::create(test_origin_signed(1)));
+		assert_ok!(Msa::create(test_origin_signed(2)));
+		assert_noop!(
+			Msa::revoke_delegation_by_delegator(test_origin_signed(1), 2),
+			Error::<Test>::DelegationNotFound
+		);
+	})
+}
+
+#[test]
+fn revoke_delegation_by_delegator_throws_error_when_delegation_already_revoked() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let provider_account = key_pair.public();
+
+		let (delegator_pair, _) = sr25519::Pair::generate();
+		let delegator_account = delegator_pair.public();
+
+		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
+
+		let provider_msa =
+			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
+
+		let (delegator_signature, add_provider_payload) =
+			create_and_sign_add_provider_payload(delegator_pair, provider_msa);
+
+		// Register provider
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(provider_account.into()),
+			Vec::from("Foo")
+		));
+
+		assert_ok!(Msa::grant_delegation(
+			RuntimeOrigin::signed(provider_account.into()),
+			delegator_account.into(),
+			delegator_signature,
+			add_provider_payload
+		));
+
+		assert_ok!(Msa::revoke_delegation_by_delegator(
+			RuntimeOrigin::signed(delegator_account.into()),
+			provider_msa
+		));
+
+		assert_noop!(
+			Msa::revoke_delegation_by_delegator(
+				RuntimeOrigin::signed(delegator_account.into()),
+				provider_msa
+			),
+			Error::<Test>::DelegationRevoked
+		);
+	});
+}
+
+/// Assert that the call to revoke a delegation is free.
+#[test]
+pub fn revoke_provider_call_has_no_cost() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let provider_account = key_pair.public();
+
+		let expiration: BlockNumber = 10;
+		let add_provider_payload = AddProvider::new(1, None, expiration);
+		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
+
+		let signature: MultiSignature = key_pair.sign(&encode_add_provider_data).into();
+
+		assert_ok!(Msa::create(test_origin_signed(1)));
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
+
+		// Register provider
+		assert_ok!(Msa::create_provider(test_origin_signed(1), Vec::from("Foo")));
+
+		assert_ok!(Msa::grant_delegation(
+			test_origin_signed(1),
+			provider_account.into(),
+			signature,
+			add_provider_payload
+		));
+
+		let call = MsaCall::<Test>::revoke_delegation_by_delegator { provider_msa_id: 2 };
+		let dispatch_info = call.get_dispatch_info();
+
+		assert_eq!(dispatch_info.pays_fee, Pays::No);
+	})
+}
+
+#[test]
+fn create_provider() {
+	new_test_ext().execute_with(|| {
+		let (_new_msa_id, key_pair) = create_account();
+
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(key_pair.public().into()),
+			Vec::from("Foo")
+		));
+	})
+}
+
+#[test]
+fn create_provider_max_size_exceeded() {
+	new_test_ext().execute_with(|| {
+		let (_new_msa_id, key_pair) = create_account();
+
+		assert_err!(
+			Msa::create_provider(
+				RuntimeOrigin::signed(key_pair.public().into()),
+				Vec::from("12345678901234567")
+			),
+			Error::<Test>::ExceedsMaxProviderNameSize
+		);
+	})
+}
+
+#[test]
+fn create_provider_duplicate() {
+	new_test_ext().execute_with(|| {
+		let (key_pair, _) = sr25519::Pair::generate();
+		let (_new_msa_id, _) =
+			Msa::create_account(key_pair.public().into(), EMPTY_FUNCTION).unwrap();
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(key_pair.public().into()),
+			Vec::from("Foo")
+		));
+
+		assert_err!(
+			Msa::create_provider(RuntimeOrigin::signed(key_pair.public().into()), Vec::from("Foo")),
+			Error::<Test>::DuplicateProviderRegistryEntry
+		)
+	})
+}
+
+pub fn set_schema_count<T: Config>(n: u16) {
+	<T>::SchemaValidator::set_schema_count(n);
+}
+
+#[test]
+pub fn valid_schema_grant() {
+	new_test_ext().execute_with(|| {
+		set_schema_count::<Test>(2);
+
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		let schema_grants = vec![1, 2];
+		assert_ok!(Msa::add_provider(provider, delegator, schema_grants));
+
+		System::set_block_number(System::block_number() + 1);
+
+		assert_ok!(Msa::ensure_valid_schema_grant(provider, delegator, 2u16, 1u64));
+	})
+}
+
+#[test]
+pub fn error_invalid_schema_id() {
+	struct TestCase<T> {
+		schema: Vec<u16>,
+		expected: T,
+	}
+	new_test_ext().execute_with(|| {
+		set_schema_count::<Test>(12);
+
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		let test_cases: [TestCase<Error<Test>>; 3] = [
+			TestCase { schema: vec![15, 16], expected: Error::<Test>::InvalidSchemaId },
+			TestCase { schema: vec![16, 17], expected: Error::<Test>::InvalidSchemaId },
+			TestCase { schema: vec![18], expected: Error::<Test>::InvalidSchemaId },
+		];
+		for tc in test_cases {
+			assert_noop!(Msa::add_provider(provider, delegator, tc.schema), tc.expected);
+		}
+	})
+}
+
+#[test]
+pub fn error_exceeding_max_schema_under_minimum_schema_grants() {
+	new_test_ext().execute_with(|| {
+		set_schema_count::<Test>(16);
+
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		assert_noop!(
+			Msa::add_provider(provider, delegator, (1..32 as u16).collect::<Vec<_>>()),
+			Error::<Test>::ExceedsMaxSchemaGrantsPerDelegation
+		);
+	})
+}
+
+#[test]
+pub fn error_not_delegated_rpc() {
+	new_test_ext().execute_with(|| {
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		assert_err!(
+			Msa::get_granted_schemas_by_msa_id(delegator, provider),
+			Error::<Test>::DelegationNotFound
+		);
+	})
+}
+
+#[test]
+pub fn error_schema_not_granted_rpc() {
+	new_test_ext().execute_with(|| {
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		assert_ok!(Msa::add_provider(provider, delegator, Vec::default()));
+		assert_err!(
+			Msa::get_granted_schemas_by_msa_id(delegator, provider),
+			Error::<Test>::SchemaNotGranted
+		);
+	})
+}
+
+#[test]
+pub fn schema_granted_success_rpc() {
+	new_test_ext().execute_with(|| {
+		set_schema_count::<Test>(2);
+
+		let provider = ProviderId(1);
+		let delegator = DelegatorId(2);
+		let schema_grants = vec![1, 2];
+		assert_ok!(Msa::add_provider(provider, delegator, schema_grants));
+		let schemas_granted = Msa::get_granted_schemas_by_msa_id(delegator, provider);
+		let expected_schemas_granted = vec![1, 2];
+		let output_schemas: Vec<SchemaId> = schemas_granted.unwrap().unwrap();
+		assert_eq!(output_schemas, expected_schemas_granted);
+	})
+}
+
+#[test]
+pub fn add_provider_expired() {
+	new_test_ext().execute_with(|| {
+		// 1. create two key pairs
+		let (provider_pair, _) = sr25519::Pair::generate();
+		let (user_pair, _) = sr25519::Pair::generate();
+
+		let provider_key = provider_pair.public();
+		let delegator_key = user_pair.public();
+
+		// 2. create provider MSA
+		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_key.into()))); // MSA = 1
+
+		// Register provider
+		assert_ok!(Msa::create_provider(
+			RuntimeOrigin::signed(provider_key.into()),
+			Vec::from("Foo")
+		));
+
+		// 3. create delegator MSA and provider to provider
+		let expiration: BlockNumber = 0;
+
+		let add_provider_payload = AddProvider::new(1u64, None, expiration);
+		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
+		let signature: MultiSignature = user_pair.sign(&encode_add_provider_data).into();
+		// 3.5 create the user's MSA + add provider as provider
+		assert_err!(
+			Msa::grant_delegation(
+				test_origin_signed(1),
+				delegator_key.into(),
+				signature,
+				add_provider_payload
+			),
+			Error::<Test>::ProofHasExpired
+		);
+	})
+}
+
+#[test]
+pub fn ensure_all_schema_ids_are_valid_errors() {
+	new_test_ext().execute_with(|| {
+		let schema_ids = vec![1];
+		assert_noop!(
+			Msa::ensure_all_schema_ids_are_valid(&schema_ids),
+			Error::<Test>::InvalidSchemaId
+		);
+
+		let schema_ids = (1..32).collect::<Vec<_>>();
+		assert_noop!(
+			Msa::ensure_all_schema_ids_are_valid(&schema_ids),
+			Error::<Test>::ExceedsMaxSchemaGrantsPerDelegation
+		);
+	})
+}
+#[test]
+pub fn ensure_all_schema_ids_are_valid_success() {
+	new_test_ext().execute_with(|| {
+		let schema_ids = vec![1];
+		set_schema_count::<Test>(1);
+
+		assert_ok!(Msa::ensure_all_schema_ids_are_valid(&schema_ids));
+	});
+}
+
+#[test]
+pub fn is_registered_provider_is_true() {
+	new_test_ext().execute_with(|| {
+		let provider = ProviderId(1);
+		let provider_name = Vec::from("frequency".as_bytes()).try_into().unwrap();
+
+		let provider_meta = ProviderRegistryEntry { provider_name };
+		ProviderToRegistryEntry::<Test>::insert(provider, provider_meta);
+
+		assert!(Msa::is_registered_provider(provider.into()));
+	});
+}
+
+#[test]
+fn delegation_default_trait_impl() {
+	new_test_ext().execute_with(|| {
+		let delegation: Delegation<
+			SchemaId,
+			<Test as frame_system::Config>::BlockNumber,
+			<Test as Config>::MaxSchemaGrantsPerDelegation,
+		> = Default::default();
+
+		let expected = Delegation {
+			schema_permissions: BoundedBTreeMap::<
+				SchemaId,
+				<Test as frame_system::Config>::BlockNumber,
+				<Test as Config>::MaxSchemaGrantsPerDelegation,
+			>::default(),
+			revoked_at: Default::default(),
+		};
+
+		assert_eq!(delegation, expected);
+	});
+}
+
+#[test]
+fn try_mutate_delegation_success() {
+	new_test_ext().execute_with(|| {
+		let delegator = DelegatorId(1);
+		let provider = ProviderId(2);
+
+		assert_ok!(Msa::try_mutate_delegation(
+			delegator,
+			provider,
+			|delegation, _is_new_provider| -> Result<(), &'static str> {
+				let schema_id = 1;
+				let _a =
+					PermittedDelegationSchemas::<Test>::try_insert_schema(delegation, schema_id);
+
+				Ok(())
+			},
+		));
+
+		assert!(Msa::get_delegation(delegator, provider).is_some());
+	});
+}
