+++ pallets/schemas/src/tests/mod.rs
@@ -1,0 +1,3 @@
+pub mod mock;
+mod other_tests;
+mod serde_tests;
